Extract ProtoStreamDataInput/Output from ProtoStreamObjectInput/Output.
