+++ clustering/marshalling/protostream/src/main/java/org/wildfly/clustering/marshalling/protostream/ProtoStreamObjectOutput.java
@@ -25,12 +25,10 @@ package org.wildfly.clustering.marshalling.protostream;
 import java.io.IOException;
 import java.io.ObjectOutput;
 import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
 
 import org.infinispan.protostream.ImmutableSerializationContext;
 import org.infinispan.protostream.ProtobufUtil;
 import org.infinispan.protostream.RawProtoStreamWriter;
-import org.infinispan.protostream.impl.RawProtoStreamWriterImpl;
 import org.wildfly.clustering.marshalling.spi.ByteBufferOutputStream;
 
 /**
@@ -38,85 +36,15 @@ import org.wildfly.clustering.marshalling.spi.ByteBufferOutputStream;
  * This implementation intentionally does not conform to the binary layout prescribed by {@link ObjectOutput}.
  * @author Paul Ferraro
  */
-public class ProtoStreamObjectOutput implements ObjectOutput {
+public class ProtoStreamObjectOutput extends ProtoStreamDataOutput implements ObjectOutput {
 
     private final ImmutableSerializationContext context;
     private final RawProtoStreamWriter writer;
 
     public ProtoStreamObjectOutput(ImmutableSerializationContext context, RawProtoStreamWriter writer) {
+        super(writer);
         this.context = context;
         this.writer = writer;
-    }
-
-    @Override
-    public void writeBoolean(boolean value) throws IOException {
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeBoolNoTag(value);
-    }
-
-    @Override
-    public void writeByte(int value) throws IOException {
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeRawByte((byte) value);
-    }
-
-    @Override
-    public void writeShort(int value) throws IOException {
-        // Write fixed length short, rather than varint
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeRawByte((byte) (0xff & (value >> 8)));
-        writer.getDelegate().writeRawByte((byte) (0xff & value));
-    }
-
-    @Override
-    public void writeChar(int value) throws IOException {
-        // Use varint encoding instead of unsigned short, as these values are most likely <= Byte.MAX_VALUE
-        this.writer.writeUInt32NoTag(value);
-    }
-
-    @Override
-    public void writeInt(int value) throws IOException {
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeSInt32NoTag(value);
-    }
-
-    @Override
-    public void writeLong(long value) throws IOException {
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeSInt64NoTag(value);
-    }
-
-    @Override
-    public void writeFloat(float value) throws IOException {
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeFloatNoTag(value);
-    }
-
-    @Override
-    public void writeDouble(double value) throws IOException {
-        RawProtoStreamWriterImpl writer = (RawProtoStreamWriterImpl) this.writer;
-        writer.getDelegate().writeDoubleNoTag(value);
-    }
-
-    @Override
-    public void writeBytes(String value) throws IOException {
-        for (int i = 0; i < value.length(); ++i) {
-            this.writeByte(value.charAt(i));
-        }
-    }
-
-    @Override
-    public void writeChars(String value) throws IOException {
-        for (int i = 0; i < value.length(); ++i) {
-            this.writeShort(value.charAt(i));
-        }
-    }
-
-    @Override
-    public void writeUTF(String value) throws IOException {
-        byte[] bytes = value.getBytes(StandardCharsets.UTF_8);
-        this.writeChar(bytes.length); // unsigned varint
-        this.write(bytes);
     }
 
     @Override
@@ -130,21 +58,6 @@ public class ProtoStreamObjectOutput implements ObjectOutput {
             this.writeChar(length); // unsigned varint
             this.write(buffer.array(), offset, length);
         }
-    }
-
-    @Override
-    public void write(int value) throws IOException {
-        this.writeByte(value);
-    }
-
-    @Override
-    public void write(byte[] bytes) throws IOException {
-        this.write(bytes, 0, bytes.length);
-    }
-
-    @Override
-    public void write(byte[] bytes, int offset, int length) throws IOException {
-        this.writer.writeRawBytes(bytes, offset, length);
     }
 
     @Override
